refactor: separate Footer component
