+++ docs/welcome-migration-plan.md
@@ -46,15 +46,16 @@
 
 ## 作業タスク
 
-1. `src/app/layout.tsx` からトップページ専用のフッター／配色を取り除き、共通設定のみを残す。同時に `src/app/page.tsx` へ移した要素を組み込み、現行トップページのデザインを保持。
+1. ~~`src/app/layout.tsx` からトップページ専用のフッター／配色を取り除き、共通設定のみを残す。同時に `src/app/page.tsx` へ移した要素を組み込み、現行トップページのデザインを保持。~~
+   - 2025-11-07: `src/app/layout.tsx` からフッターと `menuList` を削除し、`src/app/page.tsx` に移設してトップページ内で同じスタイルのフッターを表示するよう調整。
 2. `public/welcome/` ディレクトリを作成し、`temp/img` など必要画像をコピー配置 (スクリプトではなく手動で必要枚数)。
 3. `src/app/welcome/` 以下にページとコンポーネント群を実装。Tailwind クラス, データ配列, SVG アイコン整理。
 4. ヘッダーメニューの開閉ロジックを React で構築し、RWD を確認。
 5. FAQ テーブルをセマンティック HTML で再現 (Tailwind でスタイリング)。
-6. `metadata` を設定し、`noindex` を反映させる。
-7. ページ遷移後にスクロールするアンカーリンクを `id` 付与で調整。`Scroll` ボタンは hero 内に配置。
-8. `temp` ディレクトリは最終的に参照しないが、削除は次のステップで検討 (今回は保持)。
-9. ESLint/TSC チェック、`pnpm lint` or `pnpm run lint` (要確認) を実施。
+6. `metadata` を設定し、`noindex` は削除する。
+7. ページ遷移後にスクロールするアンカーリンクを `id` 付与で調整。`Scroll` ボタン(リンク)は hero 内に配置。
+8. `temp` ディレクトリは最終的に参照しないが、削除はしない。
+9. ESLint/TSC チェック、`pnpm lint` or `pnpm run lint` を実施。
 
 ## オープン課題/確認事項
